project done

+++ src/app/modules/book/book.service.ts
@@ -1,7 +1,14 @@
+import httpStatus from "http-status";
 import { IBook } from "./book.interface";
 import Book from "./book.model";
 
-const getAllBooks = async (limit?:number | undefined,searchTerm?: string, genre?: string, publicationYear?:number) => {
+
+
+const getAllBooks = async (
+  searchTerm?: string,
+  genre?: string,
+  publicationYear?: string
+) => {
   const filter: any = {};
 
   if (searchTerm) {
@@ -9,7 +16,7 @@ const getAllBooks = async (limit?:number | undefined,searchTerm?: string, genre?
     filter.$or = [
       { title: searchRegex },
       { author: searchRegex },
-      { genre: searchRegex },
+      { genre: searchRegex }
     ];
   }
 
@@ -18,30 +25,38 @@ const getAllBooks = async (limit?:number | undefined,searchTerm?: string, genre?
   }
 
   if (publicationYear) {
-    const publicationYearStart = new Date(publicationYear, 0, 1);
-    const publicationYearEnd = new Date(publicationYear, 11, 31);
-    filter.publication_date = {
-      $gte: publicationYearStart,
-      $lte: publicationYearEnd,
-    };
+    const yearNumber = parseInt(publicationYear);
+    if (!isNaN(yearNumber)) {
+      const yearStart = new Date(yearNumber, 0, 1);
+      const yearEnd = new Date(yearNumber, 11, 31, 23, 59, 59, 999);
+      filter.publication_date = {
+        $gte: yearStart,
+        $lte: yearEnd,
+      };
+    } else {
+      
+    }
   }
 
-  const result = await Book.find(filter).sort({ createdAt: -1 }).limit(limit);
+  const result = await Book.find(filter).sort({ createdAt: -1 });
   return result;
 };
+
+
 
 // get single book
 
-const getSingleBook = async (_id:string)=> {
-  const result = await Book.findById({_id})
+const getSingleBook = async (_id: string) => {
+  const result = await Book.findById({ _id });
   return result;
-}
-
+};
 
 const createBook = async (book: IBook): Promise<IBook | null> => {
   const result = await Book.create(book);
-  return book;
+  return result
 };
+
+
 
 //update book
 
@@ -49,7 +64,6 @@ const updateBook = async (
   _id: string,
   payload: Partial<IBook>
 ): Promise<IBook | null> => {
-  console.log('serviced....',payload);
   const isExist = await Book.findOne({ _id });
 
   if (!isExist) {
@@ -65,32 +79,40 @@ const updateBook = async (
 };
 
 interface reviews {
-  rating:number;
-  comment:string
+  rating: number;
+  comment: string;
 }
 
-const postReview = async (id:string,payload:reviews )=> {
+const postReview = async (id: string, payload: reviews) => {
   const isBookExit = await Book.findById(id);
-  if(!isBookExit){
-    throw new Error("Book is not exits")
+  if (!isBookExit) {
+    throw new Error("Book is not exits");
   }
-  const result = await  Book.findByIdAndUpdate(
+  const result = await Book.findByIdAndUpdate(
     id,
     { $push: { reviews: payload } },
     { new: true }
   );
 
   return result;
-}
+};
 
 // delete book
-const deleteBook = async (_id: string) => {
+const deleteBook = async (_id: string, email: string) => {
   const isExist = await Book.findOne({ _id });
   if (!isExist) {
     throw new Error("Book is not found !");
   }
-  const result = await Book.deleteOne({ _id });
-  return result;
+
+  if (isExist && isExist.user_email === email) {
+    const book = await Book.deleteOne({ _id });
+    if (book) {
+      const data = "Book deleted successfully";
+      return data;
+    }
+  }
+  const data = "You are unauthorize . cannot delete book !";
+  return data;
 };
 
 export const bookService = {
@@ -99,5 +121,5 @@ export const bookService = {
   createBook,
   updateBook,
   deleteBook,
-  postReview
+  postReview,
 };
